requirements/development.txt
-r base.txt
ipdb==0.12.2
pytest==5.1.1
pytest-django==3.5.1
pytest-sugar==0.9.2
coverage==4.5.3
pylint==2.3.1
vcrpy==2.1.0
locustio==0.12.1